requirements/tests.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/tests.in
#
attrs==19.3.0             # via pytest
cffi==1.14.0              # via cryptography
cryptography==3.2.1       # via -r requirements/tests.in
greenlet==0.4.17          # via -r requirements/tests.in
iniconfig==1.0.0          # via pytest
packaging==20.3           # via pytest
pathtools==0.1.2          # via watchdog
pluggy==0.13.1            # via pytest
psutil==5.7.0             # via pytest-xprocess
py==1.9.0                 # via pytest
pycparser==2.20           # via cffi
pyparsing==2.4.7          # via packaging
pytest-timeout==1.4.2     # via -r requirements/tests.in
pytest-xprocess==0.16.0   # via -r requirements/tests.in
pytest==6.1.2             # via -r requirements/tests.in, pytest-timeout, pytest-xprocess
six==1.14.0               # via cryptography, packaging
toml==0.10.1              # via pytest
watchdog==0.10.3          # via -r requirements/tests.in
